location-privacy-etsi/requirements.txt
lxml==4.6.4
networkx==2.6.3
numpy==1.21.4
PyYAML==6.0
sumolib==1.12.0
tqdm==4.62.3
traci==1.12.0
